requirements.txt
streamlit>=1.30
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0
numpy>=1.24
